refactor(ai-plugins): migrate ai transformer plugins to new framework
