+++ spec/03-plugins/39-ai-request-transformer/02-integration_spec.lua
@@ -145,7 +145,7 @@ local _EXPECTED_CHAT_STATS = {
     meta = {
       plugin_id = '71083e79-4921-4f9f-97a4-ee7810b6cd8a',
       provider_name = 'openai',
-      request_model = 'gpt-4',
+      request_model = 'UNSPECIFIED',
       response_model = 'gpt-3.5-turbo-0613',
       llm_latency = 1
     },
@@ -166,7 +166,7 @@ local SYSTEM_PROMPT = "You are a mathematician. "
 
 local client
 
-for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
+for _, strategy in helpers.all_strategies() do
   describe(PLUGIN_NAME .. ": (access) [#" .. strategy .. "]", function()
 
     lazy_setup(function()
@@ -404,5 +404,4 @@ for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
       end)
     end)
   end)
-end
 end
